// src/Api/Model/Artist/SpotifyArtist.ts
export type SpotifyArtist = {
  readonly id: string
  readonly name: string
  readonly uri: string
  readonly url: string
  readonly genres: string[]
  readonly popularity: number
  readonly follower: number
  readonly thumbnailImage: string
  readonly smallImage: string
  readonly mediumImage: string
  readonly largeImage: string
}